+++ kumihan_formatter/parsers/list/__init__.py
@@ -3,9 +3,12 @@
 責任分離により以下の構造で分割:
 - list_parser.py: メインクラスとパブリックインターフェース
 - list_handlers.py: リスト処理・分析ロジック
+- list_item_handlers.py: 個別アイテム処理ハンドラ
 - list_utils.py: ユーティリティ関数群
 """
 
 from .list_parser import UnifiedListParser
+from .list_handlers import ListHandler
+from .list_item_handlers import ListItemHandler
 
-__all__ = ["UnifiedListParser"]
+__all__ = ["UnifiedListParser", "ListHandler", "ListItemHandler"]
